refactor: reorganizes file structure
