+++ src/components/RenderForm.jsx
@@ -2,7 +2,7 @@ import { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
 import urlBuilder from '../utilities/utils';
 import '../styles/RenderForm.css';
-import DataTable from './DataTable';
+import DataTable from './DataTable/DataTable';
 
 function RenderForm({ userInput }) {
   /**
